// src/modules/permission/permission.enum.ts
export enum Permission {
  Assignments = 'assignments',
  ESignature = 'eSignature',
  Timesheets = 'timesheets',
  Contracts = 'contracts',
  Invoices = 'invoices',
  StaffingRequests = 'staffingRequests',
  TimeSheetApprover = 'timeSheetApprover',
  ReportTo = 'reportTo',
  BillTo = 'billTo',
  AP = 'ap',
  Main = 'main',
}

export enum PermissionAction {
  Write = 'write',
  Read = 'read',
}